feat: add start_state for selecting game state move
Add start_state subroutine which manages the player's choice for move.
It allows the player the select rock, paper, or scissors. The choice is
stored in a labeled register which holds the local player's game state.
During selection, the currently selected choice is displayed on the
second line of the LCD

Add const_copy_prog_to_data_16 macro which simplifies the calling of
this subroutine. Changes multi-line call to single line call allowing
the code to be denser and more abstracted.